Added Abstract, Conclusion; Updated DEP -> Sampled DEP
Wrote paper abstract and conclusion, finishing what remained of the paper.  Paper is effectively complete.

Also, after discussion with MacKenzie, I have a different understanding of what the DEP is compared to when I originally wrote the paper.  Updated section III(b) to include these changes (DEP is only equivalent when Omega is infinite, and the computationally usable version is a sampled form of the DEP), and updated references to the DEP to sampled DEP.